rest_api/requirements.txt
Django==2.0.1
djangorestframework==3.9.4
docker==4.0.1
